requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
openai>=1.3.0
twilio>=8.10.0
requests>=2.31.0
pydantic>=2.5.0
python-multipart>=0.0.6
pytest>=7.4.3
pytest-asyncio>=0.23.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
sentry-sdk>=1.35.0
click>=8.1.7
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0 